Move Parent and Child fixture classes to conftest
