+++ skbase/tests/mock_package/test_mock_package.py
@@ -47,39 +47,6 @@ class CompositionDummy(BaseObject):
         return [params1, params2]
 
 
-# Fixture class for testing tag system
-class Parent(BaseObject):
-    """Parent class to test BaseObject's usage."""
-
-    _tags = {"A": "1", "B": 2, "C": 1234, "3": "D"}
-
-    def __init__(self, a="something", b=7, c=None):
-        self.a = a
-        self.b = b
-        self.c = c
-        super().__init__()
-
-    def some_method(self):
-        """To be implemented by child class."""
-        pass
-
-
-# Fixture class for testing tag system, child overrides tags
-class Child(Parent):
-    """Child class that is child of FixtureClassParent."""
-
-    _tags = {"A": 42, "3": "E"}
-    __author__ = ["fkiraly", "RNKuhns"]
-
-    def some_method(self):
-        """Child class' implementation."""
-        pass
-
-    def some_other_method(self):
-        """To be implemented in the child class."""
-        pass
-
-
 class InheritsFromBaseObject(BaseObject):
     """A class inheriting from BaseObject."""
 
@@ -101,9 +68,7 @@ class _NonPublicClass(BaseObject):
 
 MOCK_PACKAGE_OBJECTS = [
     AnotherClass,
-    Child,
     CompositionDummy,
     InheritsFromBaseObject,
-    Parent,
     _NonPublicClass,
 ]
